📚 design: First cut of Teleport challenge design (#1)
Add the first cut of the design doc addressing the Teleport interview
challenge. Add the challenge document itself alongside it.

Also present is the first cut of the protobuf specification for the
`JobExecutor` service. It is subject to revision as the challenge is
implemented.

Pull-Request: https://github.com/camh-/tpchallenge/pull/1