Add protoc hermit packages for generating Go code
Add packages with hermit for running `protoc` to generate Go bindings
for protobuf and grpc:

    hermit install protoc-3.20.1 protoc-gen-go protoc-gen-go-grpc

Note: Hours before this commit, the hermit package repository updated
the `protoc` manifest for version 21.0. It looks like the versioning
scheme for protoc may have changed as well as the packaging format, an
at the moment 21.0 fails to install with hermit, hence the version
specifier of `3.20.1` on the hermit install command for protoc.