cli: Implement the CLI running against a fake server
Implement the CLI part of jobber implementing the client commands:

    jobber run [-d] command [args...]
    jobber stop [-c] job-id
    jobber status job-id
    jobber list [-t] [-a]
    jobber logs [-f] job-id

Also implement `jobber serve` that serve a fake `JobExecutor` service
that is also used for the client cli tests. The fake service will be
swapped out for the real one once it is implement, but as is, it gives a
way to run the server and client and try it out on the command line.

Some new targets have been added to the Makefile. Run `make help` to see
the list of targets and what they do. Added here are:

    make build
    make test
    make clean