job: Add implementation of containerised job running
Add an implementation of running jobs in a container, fully implementing
the `JobExecutor` service.

This implements the `Job` and `job.Tracker` of the design doc, although
the tracker and job have been unified into a single package.

Container execution includes the cgroups and namespaces specified in the
design doc.

Currently unimplemented is proper authorization - just a single static
user is hard-coded, although the tracker does use that to implement
authorization. Actual authentication will be implemented with the TLS
implementation where the user can be extracted from the client
certificate.

This needs some tests :( - they will come soon.