Add certstrap tool for creating certs
Add the certstrap tool for creating CA, server and user certificates for
use with jobber.

Unfortunately certstrap has to be installed from source as the version
in hermit is 3 years old (as the certstrap repo has not released a
binary or tag in that long), so we just build from @master. For this, we
add to the environment using hermit to set GOBIN to out/bin and put that
first on the path.

Add make targets for creating the certs:

    make certs/ca.crt  # make the CA certs (will make .key file too)
    make certs/server.crt  # make the server certs (+key)
    make certs/camh.crt  # make a user/client cert (+key)

    make default-user-cert

This will make a user/client cert for the user named in `$USER` and
symlink `certs/user.crt` (+key) to it. This gives a standard name for
the code to use for a default user cert.

Ignore the `certs/` directory as these should not be committed (nor do
they need to be - we just generate certs fresh to use/test jobber).