tls: Use TLS in client and server
Add options for certs, keys and CAs in the client and serve with default
values - `certs/server.{crt,key}`, `certs/user.{crt,key}` and
`certs/ca,crt`. See the previous commit message for details on creating
the certs.

Change the default port to 8443 from 8080 to indicate TLS is required.

The error messages that come out when cert validation fails,
particularly from bad client certs, are quite terrible. I'll need to
figure out how to interpret these for better error messages,

The CN is the user cert is not propagated into the context for
`job.Tracker` to use. That comes next.

Update the tests to use test certs created by the makefile specifically
for the tests. We also create a bad user and server cert (from a
different CA) so we can test that certs fail when they should.