auth: Add admin users
Add admin users to the tracker to allow users in the list of admins to
operate on jobs that are not their own. The list is populated from the
`--admin` flag to `jobber serve`.

Make `make run-server` add the user running it to the list of admins.

Different users can be tested by running multiple commands with
different user certs:

    USER=eve make default-user-cert
    ./out/jobber run -d /bin/sleep 100
    ./out/jobber run -d /bin/echo hello world
    make default-user-cert  # run with your own user, admin
    ./out/jobber run -d /bin/false
    ./out/jobber run -d /bin/sleep 100
    ./out/jobber list
    ./out/jobber list -a
    ./out/jobber list -ac

`jobber list` makes it easy to see multiple jobs at once, but you can
also try to stop a job as different non-admin user to get an
authorization failure error.