shutdown: Add a shutdown command to API and CLI
Add a `Shutdown()` method to the gRPC API and add a CLI subcommand to
the client to call it. Only admin users can issue a shutdown command.

Shutdown stops all running jobs immediately and returns the number of
jobs stopped. It gracefully stops the gRPC server allowing existing
commands to complete. Once the shutdown is initiated, no new jobs may be
started.

It is conceivable that the `Shutdown()` API method should perhaps be in
a separate admin API, rather than in the `JobExecutor` API. A service
implementing multiple services only needs one shutdown method.